+++ plugins/claim_mapper/_abstract.rb
@@ -6,7 +6,18 @@ require 'abstraction'
 class ClaimMapper
   abstract
 
-  def self.map_claims(claims, provider)
+  # Used to map additional claims into the access token
+  def map_to_access_token(_client, _user, _scopes, _requested_claims, _resources)
+    {}
+  end
+
+  # Used to map additional claims into the id token
+  def map_to_id_token(_client, _user, _scopes, _requested_claims)
+    {}
+  end
+
+  # Used to generate external users
+  def map_from_provider(claims, provider)
     raise NotImplementedError
   end
 end
