+++ tests/test_admin_api.rb
@@ -4,7 +4,7 @@ require 'rack/test'
 require 'webrick/https'
 require_relative 'config_testsetup'
 require_relative '../omejdn'
-require_relative '../lib/token_helper'
+require_relative '../lib/token'
 
 class AdminApiTest < Test::Unit::TestCase
   include Rack::Test::Methods
@@ -17,8 +17,8 @@ class AdminApiTest < Test::Unit::TestCase
     TestSetup.setup
     
     client = Client.find_by_id 'testClient'
-    @token = TokenHelper.build_access_token client, nil, ['omejdn:admin'], {}, TestSetup.config['host']+"/api"
-    @insufficient_token = TokenHelper.build_access_token client, nil, ['omejdn:write'], {}, "test"
+    @token = Token.access_token client, nil, ['omejdn:admin'], {}, TestSetup.config['host']+"/api"
+    @insufficient_token = Token.access_token client, nil, ['omejdn:write'], {}, "test"
     @testCertificate = File.read './tests/test_resources/testClient.pem'
   end
