+++ tests/test_oauth2.rb
@@ -4,7 +4,7 @@ require 'rack/test'
 require 'webrick/https'
 require_relative 'config_testsetup'
 require_relative '../omejdn'
-require_relative '../lib/token_helper'
+require_relative '../lib/token'
 
 class OAuth2Test < Test::Unit::TestCase
   include Rack::Test::Methods
